+++ larpmanager/templates/larpmanager/exe/users/archive_mail.html
@@ -1,0 +1,34 @@
+{% extends 'utils.html' %}
+{% load i18n %}
+{% block title %}
+    {% trans "Archive mail" %} - {% trans "Organization" %}
+{% endblock %}
+{% block info %}
+    {% trans "This page shows all emails, both queued and sent." %}
+{% endblock %}
+{% block content %}
+    {% include "elements/header/exe.html" %}
+    {% include "elements/paginate.html" %}
+    <table id="inflows" class="tablesorter mob">
+        <thead>
+            <tr>
+                <th>{% trans "Run" %}</th>
+                <th>{% trans "Recipient" %}</th>
+                <th>{% trans "Subject" %}</th>
+                <th>{% trans "Body" %}</th>
+                <th>{% trans "Sent" %}</th>
+            </tr>
+        </thead>
+        <tbody>
+            {% for el in list %}
+                <tr id="{{ el.id }}">
+                    <td>{% if el.run %}{{ el.run }}{% endif %}</td>
+                    <td>{{ el.recipient }}</td>
+                    <td>{{ el.subj }}</td>
+                    <td>{{ el.body | striptags | truncatechars:100 }}</td>
+                    <td>{% if el.sent %}{{ el.sent | date:"d/m/Y" }}{% endif %}</td>
+                </tr>
+            {% endfor %}
+        </tbody>
+    </table>
+{% endblock %}
